Removing SpApiInterface and letting SpApi become a more obvious package wrapper; changing tests to match its intended use more closely
