+++ src/Services/Authenticator/ClientAuthenticator.php
@@ -5,7 +5,7 @@ namespace Glue\SpApi\OpenAPI\Services\Authenticator;
 use Glue\SpApi\OpenAPI\Configuration\SpApiConfig;
 use Glue\SpApi\OpenAPI\Exceptions\LwaAccessTokenException;
 use Glue\SpApi\OpenAPI\Exceptions\RestrictedDataTokenException;
-use Glue\SpApi\OpenAPI\Middleware\AwsSignatureV4Middleware;
+use Glue\SpApi\OpenAPI\Middleware\Guzzle\AwsSignatureV4Middleware;
 use Glue\SpApi\OpenAPI\Services\Lwa\LwaServiceInterface;
 use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
