+++ tests/Clients/CatalogItemsV0/Api/CatalogApiTest.php
@@ -4,27 +4,20 @@ namespace Tests\Clients\CatalogItemsV0\Api;
 
 use Glue\SpApi\OpenAPI\Clients\CatalogItemsV0\Model\ListCatalogItemsResponse;
 use Glue\SpApi\OpenAPI\Clients\CatalogItemsV0\Model\ListMatchingItemsResponse;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class CatalogApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_listCatalogItems()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->catalogItemsV0()
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->catalogItemsV0()
                 ->listCatalogItemsWithHttpInfo('TEST_CASE_200', null, null, 'SKU_200');
         });
