+++ tests/Clients/CatalogItemsV20201201/Api/CatalogApiTest.php
@@ -4,27 +4,20 @@ namespace Tests\Clients\CatalogItemsV20201201\Api;
 
 use Glue\SpApi\OpenAPI\Clients\CatalogItemsV20201201\Model\Item;
 use Glue\SpApi\OpenAPI\Clients\CatalogItemsV20201201\Model\ItemSearchResults;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class CatalogApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_searchCatalogItems()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->catalogItemsV20201201()->searchCatalogItemsWithHttpInfo(
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->catalogItemsV20201201()->searchCatalogItemsWithHttpInfo(
                 ['red', 'polo', 'shirt'],
                 ['ATVPDKIKX0DER'],
                 'summaries'
