+++ tests/Clients/FbaSmallAndLightV1/Api/SmallAndLightApiTest.php
@@ -4,27 +4,20 @@ namespace Tests\Clients\FbaSmallAndLightV1\Api;
 
 use Glue\SpApi\OpenAPI\Clients\FbaSmallAndLightV1\Model\SmallAndLightEnrollment;
 use Glue\SpApi\OpenAPI\Clients\FbaSmallAndLightV1\Model\SmallAndLightEnrollmentStatus;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class SmallAndLightApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getSmallAndLightEnrollmentBySellerSKU()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->fbaSmallAndLightV1()
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->fbaSmallAndLightV1()
                 ->getSmallAndLightEnrollmentBySellerSKUWithHttpInfo(
                     'SKU_ENROLLED_IN_SMALL_AND_LIGHT',
                     ['ATVPDKIKX0DER']
