+++ tests/Clients/FinancesV0/Api/FinancesApiTest.php
@@ -5,27 +5,20 @@ namespace Tests\Clients\FinancesV0\Api;
 use Glue\SpApi\OpenAPI\Clients\FinancesV0\Model\FinancialEventGroup;
 use Glue\SpApi\OpenAPI\Clients\FinancesV0\Model\ListFinancialEventGroupsPayload;
 use Glue\SpApi\OpenAPI\Clients\FinancesV0\Model\ListFinancialEventGroupsResponse;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class FinancesApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_listFinancialEventGroups()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->financesV0()->listFinancialEventGroupsWithHttpInfo(
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->financesV0()->listFinancialEventGroupsWithHttpInfo(
                 1,
                 '2019-10-31',
                 '2019-10-13'
