+++ tests/Clients/FulfillmentOutboundV20200701/Api/FbaOutboundApiTest.php
@@ -9,32 +9,25 @@ use Glue\SpApi\OpenAPI\Clients\FulfillmentOutboundV20200701\Model\GetFulfillment
 use Glue\SpApi\OpenAPI\Clients\FulfillmentOutboundV20200701\Model\GetFulfillmentPreviewResponse;
 use Glue\SpApi\OpenAPI\Clients\FulfillmentOutboundV20200701\Model\GetFulfillmentPreviewResult;
 use Glue\SpApi\OpenAPI\Clients\FulfillmentOutboundV20200701\Model\Money;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class FbaOutboundApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getFulfillmentPreview()
     {
         $this->markTestIncomplete('Should revisit after regenerating clients via openapi-generator-cli option --additional-properties=enumUnknownDefaultCase=true');
 
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->fulfillmentOutboundV20200701()
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->fulfillmentOutboundV20200701()
                 ->getFulfillmentPreviewWithHttpInfo(
                     new GetFulfillmentPreviewRequest([
-                        'marketplaceId' => $this->spApi->getSpApiConfig()->defaultMarketplaceId,
+                        'marketplaceId' => $this->sp_api()->getSpApiConfig()->defaultMarketplaceId,
                         'address'       => new Address([
                             'name'          => 'Walts TV and Appliance',
                             'addressLine1'  => '1746 W Ruby Dr',
