+++ tests/Clients/NotificationsV1/Api/NotificationsApiTest.php
@@ -4,27 +4,20 @@ namespace Tests\Clients\NotificationsV1\Api;
 
 use Glue\SpApi\OpenAPI\Clients\NotificationsV1\Model\GetSubscriptionResponse;
 use Glue\SpApi\OpenAPI\Clients\NotificationsV1\Model\Subscription;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class NotificationsApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getSubscription()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->notificationsV1()
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->notificationsV1()
                 ->getSubscriptionWithHttpInfo('LISTINGS_ITEM_ISSUES_CHANGE');
         });
