+++ tests/Clients/OrdersV0/Api/OrdersV0ApiTest.php
@@ -6,28 +6,21 @@ use Glue\SpApi\OpenAPI\Clients\OrdersV0\Model\GetOrderResponse;
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Model\GetOrdersResponse;
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Model\Order;
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Model\OrdersList;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class OrdersV0ApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getOrders()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->ordersV0()->getOrdersWithHttpInfo(
-                [$this->spApi->getSpApiConfig()->defaultMarketplaceId],
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->ordersV0()->getOrdersWithHttpInfo(
+                [$this->sp_api()->getSpApiConfig()->defaultMarketplaceId],
                 // Using this specific string value as a quirky requirement of the sandbox API (see models/ordersV0.json)
                 'TEST_CASE_200'
             );
@@ -47,8 +40,8 @@ class OrdersV0ApiTest extends TestCase
 
     public function test_getOrder()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->ordersV0()->getOrderWithHttpInfo(
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->ordersV0()->getOrderWithHttpInfo(
                 // Using this specific string value as a quirky requirement of the sandbox API (see models/ordersV0.json)
                 'TEST_CASE_200'
             );
