+++ tests/Clients/OrdersV0/Api/ShipmentApiTest.php
@@ -4,30 +4,23 @@ namespace Tests\Clients\OrdersV0\Api;
 
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Model\ShipmentStatus;
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Model\UpdateShipmentStatusRequest;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class ShipmentApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_updateShipmentStatus()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->ordersV0Shipment()->updateShipmentStatusWithHttpInfo(
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->ordersV0Shipment()->updateShipmentStatusWithHttpInfo(
                 'testOrder123',
                 new UpdateShipmentStatusRequest([
-                    'marketplaceId'  => $this->spApi->getSpApiConfig()->defaultMarketplaceId,
+                    'marketplaceId'  => $this->sp_api()->getSpApiConfig()->defaultMarketplaceId,
                     'shipmentStatus' => ShipmentStatus::READY_FOR_PICKUP,
                 ])
             );
