+++ tests/Clients/ReplenishmentV20221107/Api/SellingpartnersApiTest.php
@@ -10,27 +10,20 @@ use Glue\SpApi\OpenAPI\Clients\ReplenishmentV20221107\Model\Metric;
 use Glue\SpApi\OpenAPI\Clients\ReplenishmentV20221107\Model\ProgramType;
 use Glue\SpApi\OpenAPI\Clients\ReplenishmentV20221107\Model\TimeInterval;
 use Glue\SpApi\OpenAPI\Clients\ReplenishmentV20221107\Model\TimePeriodType;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class SellingpartnersApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getSellingPartnerMetrics()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->replenishmentV20221107Sellingpartners()
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->replenishmentV20221107Sellingpartners()
                 ->getSellingPartnerMetricsWithHttpInfo(
                     new GetSellingPartnerMetricsRequest([
                         'aggregationFrequency' => AggregationFrequency::YEAR,
