+++ tests/Clients/ReportsV20210630/Api/ReportsApiTest.php
@@ -4,31 +4,24 @@ namespace Tests\Clients\ReportsV20210630\Api;
 
 use Glue\SpApi\OpenAPI\Clients\ReportsV20210630\Model\Report;
 use Glue\SpApi\OpenAPI\Clients\ReportsV20210630\Model\GetReportsResponse;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class ReportsApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getReports()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->reportsV20210630()->getReportsWithHttpInfo(
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->reportsV20210630()->getReportsWithHttpInfo(
                 // Specific values come from the sandbox spec in models/reports_2021-06-30.json
                 ['FEE_DISCOUNTS_REPORT', 'GET_AFN_INVENTORY_DATA'],
                 ['IN_QUEUE', 'IN_PROGRESS'],
-                [$this->spApi->getSpApiConfig()->defaultMarketplaceId],
+                [$this->sp_api()->getSpApiConfig()->defaultMarketplaceId],
                 10
             );
         });
