+++ tests/Clients/SellersV1/Api/SellersApiTest.php
@@ -4,27 +4,20 @@ namespace Tests\Clients\SellersV1\Api;
 
 use Glue\SpApi\OpenAPI\Clients\SellersV1\Model\GetMarketplaceParticipationsResponse;
 use Glue\SpApi\OpenAPI\Clients\SellersV1\Model\MarketplaceParticipation;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class SellersApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getMarketplaceParticipations()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->sellersV1()->getMarketplaceParticipationsWithHttpInfo();
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->sellersV1()->getMarketplaceParticipationsWithHttpInfo();
         });
 
         /**
