+++ tests/Clients/VendorDirectFulfillmentOrdersV20211228/Api/VendorOrdersApiTest.php
@@ -4,28 +4,21 @@ namespace Tests\Clients\VendorDirectFulfillmentOrdersV20211228\Api;
 
 use Glue\SpApi\OpenAPI\Clients\VendorDirectFulfillmentOrdersV20211228\Model\Order;
 use Glue\SpApi\OpenAPI\Clients\VendorDirectFulfillmentOrdersV20211228\Model\OrderList;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class VendorOrdersApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getOrders()
     {
         $result = $this->tryButSkipIfUnauthorized(function () {
-            return $this->spApi->execute(function () {
-                return $this->spApi->vendorDirectFulfillmentOrdersV20211228()->getOrdersWithHttpInfo(
+            return $this->sp_api()->execute(function () {
+                return $this->sp_api()->vendorDirectFulfillmentOrdersV20211228()->getOrdersWithHttpInfo(
                     '2020-02-15T14:00:00-08:00',
                     '2020-02-20T00:00:00-08:00',
                     null,
