+++ tests/Clients/VendorDirectFulfillmentShippingV20211228/Api/VendorShippingApiTest.php
@@ -3,28 +3,21 @@
 namespace Tests\Clients\VendorDirectFulfillmentShippingV20211228\Api;
 
 use Glue\SpApi\OpenAPI\Clients\VendorDirectFulfillmentShippingV20211228\Model\PackingSlipList;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class VendorShippingApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getPackingSlips()
     {
         $result = $this->tryButSkipIfUnauthorized(function () {
-            return $this->spApi->execute(function () {
-                return $this->spApi->vendorDirectFulfillmentShippingV20211228()
+            return $this->sp_api()->execute(function () {
+                return $this->sp_api()->vendorDirectFulfillmentShippingV20211228()
                     ->getPackingSlipsWithHttpInfo(
                         '2020-02-15T14:00:00-08:00',
                         '2020-02-20T00:00:00-08:00',
