+++ tests/Clients/VendorTransactionStatusV1/Api/VendorTransactionApiTest.php
@@ -4,27 +4,20 @@ namespace Tests\Clients\VendorTransactionStatusV1\Api;
 
 use Glue\SpApi\OpenAPI\Clients\VendorTransactionStatusV1\Model\GetTransactionResponse;
 use Glue\SpApi\OpenAPI\Clients\VendorTransactionStatusV1\Model\TransactionStatus;
-use Glue\SpApi\OpenAPI\Container\SpApi;
 use Tests\TestCase;
 
 class VendorTransactionApiTest extends TestCase
 {
-    /**
-     * @var SpApi
-     */
-    public $spApi;
-
     // TODO: This will need to be changed to `public function setUp(): void` after upgrading.
     public function setUp()
     {
         parent::setup();
-        $this->spApi = $this->buildSpApiContainer();
     }
 
     public function test_getTransaction()
     {
-        $result = $this->spApi->execute(function () {
-            return $this->spApi->vendorTransactionStatusV1()
+        $result = $this->sp_api()->execute(function () {
+            return $this->sp_api()->vendorTransactionStatusV1()
                 ->getTransactionWithHttpInfo('20190904190535-eef8cad8-418e-4ed3-ac72-789e2ee6214a');
         });
